refactor(cloud-workers): collapse duplicated workspace conflict and quiescence paths
1. Centralize conflict projection, supersession, cleanup, and settlement across live turns, recovery, and reclaim while keeping resume and reclaim terminal states explicit.
2. Reuse an entries-only changed-path helper for file and symlink staging.
3. Reuse the unchanged-branch reconcile preflight when no filesystem mutation occurred.
4. Move all quiescence scripts into one owner and share their process and lease engines. Deliberate micro-change: give quiesce processes() the same 2000 ms ps timeout added to renew by #112186.
5. Replace active-turn-claim message matching with ActiveTurnClaimError while preserving the message.
6. Document each staged-finalize stability and active-lease fence by its protected race.
7. Remove the unshipped locale-collated accepted-ref fallback and locale argv plumbing. Writer 8631832048cfd5f08f33e1899a373914e515c579 was replaced by canonical collation in 792f5b7b74e433f9471eed5abcb0fc334d28bdeb; git tag --contains returned no release tags for either commit.
