+++ src/gateway/worker-environments/workspace-quiescence-scripts.ts
@@ -1,0 +1,403 @@
+const REMOTE_QUIESCENCE_PS_JS = String.raw`function processes() {
+  const output = childProcess.execFileSync("ps", ["-axo", "pid=,ppid=,uid=,stat=,lstart="], {
+    encoding: "utf8",
+    maxBuffer: 4 * 1024 * 1024,
+    timeout: 2000,
+  });
+  const rows = new Map();
+  for (const line of output.split("\n")) {
+    const match = line.trim().match(/^(\d+)\s+(\d+)\s+(\d+)\s+(\S+)\s+(.+)$/);
+    if (!match) continue;
+    rows.set(Number(match[1]), {
+      ppid: Number(match[2]),
+      uid: Number(match[3]),
+      state: match[4],
+      start: match[5],
+    });
+  }
+  return rows;
+}
+function ancestors(rows) {
+  const result = new Set();
+  let pid = process.pid;
+  while (pid > 0 && !result.has(pid)) {
+    result.add(pid);
+    pid = rows.get(pid)?.ppid || 0;
+  }
+  return result;
+}
+function processIdentity(pid) {
+  try {
+    const start = require("node:child_process").execFileSync("ps", ["-o", "lstart=", "-p", String(pid)], {
+      encoding: "utf8",
+      maxBuffer: 4096,
+    }).trim();
+    return start || null;
+  } catch (error) {
+    if (error && error.status === 1) return null;
+    throw error;
+  }
+}
+function processStatus(pid) {
+  try {
+    const output = childProcess.execFileSync("ps", ["-o", "stat=,lstart=", "-p", String(pid)], { encoding: "utf8", maxBuffer: 4096, timeout: 2000 }).trim();
+    const match = /^(\S+)\s+(.+)$/u.exec(output);
+    return match ? { state: match[1], start: match[2] } : null;
+  } catch (error) {
+    if (error && error.status === 1) return null;
+    throw error;
+  }
+}
+function quiescenceCandidates(rows, expectedUid, excludedPids, frozen) {
+  const preserved = ancestors(rows);
+  return [...rows.entries()].filter(
+    ([pid, row]) =>
+      row.uid === expectedUid &&
+      !preserved.has(pid) &&
+      row.ppid !== process.pid &&
+      !excludedPids.has(pid) &&
+      (!frozen || !frozen.has(pid)) &&
+      !row.state.startsWith("T") &&
+      !row.state.startsWith("Z") &&
+      !row.state.startsWith("X"),
+  );
+}`;
+
+const REMOTE_QUIESCENCE_LEASE_JS = String.raw`function validProcessReference(value) {
+  return value && Number.isSafeInteger(value.pid) && value.pid > 0 && typeof value.start === "string" && value.start.length > 0 && value.start.length <= 128;
+}
+function parseLease(raw, expectedNonce, options = {}) {
+  const lease = JSON.parse(raw);
+  if (
+    !lease ||
+    lease.version !== 1 ||
+    lease.nonce !== expectedNonce ||
+    !Array.isArray(lease.processes) ||
+    lease.processes.length > 4096 ||
+    lease.processes.some((entry) => !validProcessReference(entry)) ||
+    (lease.watchdog !== null && !validProcessReference(lease.watchdog)) ||
+    (options.requireWatchdog && lease.watchdog === null) ||
+    !Number.isSafeInteger(lease.expiresAtMs) ||
+    lease.expiresAtMs < 1 ||
+    (options.minimumRemainingMs && lease.expiresAtMs - Date.now() < options.minimumRemainingMs)
+  ) {
+    throw new Error(options.errorMessage || "invalid workspace quiescence lease");
+  }
+  return lease;
+}
+function persistLease(targetPath, lease, verifyCurrent) {
+  if (verifyCurrent) verifyCurrent(JSON.parse(fs.readFileSync(targetPath, "utf8")));
+  const temporary = targetPath + "." + process.pid + "." + crypto.randomBytes(8).toString("hex");
+  fs.writeFileSync(temporary, JSON.stringify(lease), { mode: 0o600, flag: "wx" });
+  fs.renameSync(temporary, targetPath);
+}`;
+
+export const REMOTE_WORKSPACE_QUIESCE_JS = String.raw`const childProcess = require("node:child_process");
+const crypto = require("node:crypto");
+const fs = require("node:fs");
+const os = require("node:os");
+const path = require("node:path");
+const root = fs.realpathSync(process.argv[1]);
+if (typeof process.getuid !== "function") throw new Error("workspace quiescence requires POSIX");
+const uid = process.getuid();
+if (uid === 0) throw new Error("workspace quiescence refuses root-owned worker sessions");
+const sleeper = new Int32Array(new SharedArrayBuffer(4));
+const leaseDirectory = path.join(os.homedir(), ".openclaw-worker", "quiescence");
+fs.mkdirSync(leaseDirectory, { recursive: true, mode: 0o700 });
+fs.chmodSync(leaseDirectory, 0o700);
+const workspaceKey = crypto.createHash("sha256").update(root).digest("hex");
+const nonce = crypto.randomBytes(16).toString("hex");
+const leasePath = path.join(leaseDirectory, workspaceKey + "." + nonce + ".json");
+const watchdogTimeoutMs = Number(process.argv[2] || 12 * 60 * 1000);
+if (!Number.isSafeInteger(watchdogTimeoutMs) || watchdogTimeoutMs < 1) throw new Error("invalid watchdog timeout");
+${REMOTE_QUIESCENCE_PS_JS}
+${REMOTE_QUIESCENCE_LEASE_JS}
+const frozen = new Map();
+let watchdogReference = null;
+function writeLease(expiresAtMs = Date.now() + watchdogTimeoutMs) {
+  persistLease(leasePath, {
+    version: 1,
+    nonce,
+    processes: [...frozen].map(([pid, start]) => ({ pid, start })),
+    watchdog: watchdogReference,
+    expiresAtMs,
+  });
+}
+function resumeProcesses(entries) {
+  for (const entry of entries) {
+    if (processIdentity(entry.pid) !== entry.start) continue;
+    try {
+      process.kill(entry.pid, "SIGCONT");
+    } catch (error) {
+      if (!error || error.code !== "ESRCH") throw error;
+    }
+  }
+}
+const orphanNames = fs.readdirSync(leaseDirectory).filter((name) =>
+  name.startsWith(workspaceKey + ".") && name.endsWith(".json"),
+);
+if (orphanNames.length > 16) throw new Error("too many workspace quiescence leases");
+for (const name of orphanNames) {
+  const match = name.match(/^[a-f0-9]{64}\.([a-f0-9]{32})\.json$/);
+  if (!match) continue;
+  const orphanPath = path.join(leaseDirectory, name);
+  const lease = parseLease(fs.readFileSync(orphanPath, "utf8"), match[1]);
+  if (lease.watchdog !== null && processIdentity(lease.watchdog.pid) === lease.watchdog.start) {
+    try { process.kill(lease.watchdog.pid, "SIGTERM"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
+  }
+  resumeProcesses(lease.processes);
+  fs.unlinkSync(orphanPath);
+}
+writeLease();
+const watchdog = childProcess.spawn(
+  process.execPath,
+  ["-e", processIdentity.toString() + "\n(" + watchdogMain.toString() + ")(process.argv[1], process.argv[2])", leasePath, nonce],
+  { detached: true, stdio: "ignore" },
+);
+watchdog.unref();
+if (!Number.isSafeInteger(watchdog.pid) || watchdog.pid < 1) {
+  fs.unlinkSync(leasePath);
+  throw new Error("workspace quiescence watchdog did not start");
+}
+let watchdogStart = null;
+for (let attempt = 0; attempt < 100 && !watchdogStart; attempt += 1) {
+  watchdogStart = processIdentity(watchdog.pid);
+  if (!watchdogStart) Atomics.wait(sleeper, 0, 0, 10);
+}
+if (!watchdogStart) {
+  try { process.kill(watchdog.pid, "SIGTERM"); } catch {}
+  fs.unlinkSync(leasePath);
+  throw new Error("workspace quiescence watchdog identity was not observable");
+}
+watchdogReference = { pid: watchdog.pid, start: watchdogStart };
+writeLease();
+let quietScans = 0;
+try {
+  for (let attempt = 0; attempt < 250 && quietScans < 3; attempt += 1) {
+    const candidates = quiescenceCandidates(
+      processes(),
+      uid,
+      new Set([watchdog.pid]),
+      frozen,
+    );
+    if (candidates.length + frozen.size > 4096) {
+      throw new Error("too many worker processes to quiesce safely");
+    }
+    for (const [pid, row] of candidates) {
+      try {
+        frozen.set(pid, row.start);
+        writeLease();
+        if (processIdentity(pid) !== row.start) {
+          frozen.delete(pid);
+          writeLease();
+          continue;
+        }
+        process.kill(pid, "SIGSTOP");
+      } catch (error) {
+        if (!error || error.code !== "ESRCH") throw error;
+      }
+    }
+    Atomics.wait(sleeper, 0, 0, 20);
+    const writable = quiescenceCandidates(
+      processes(),
+      uid,
+      new Set([watchdog.pid]),
+    ).length > 0;
+    quietScans = writable ? 0 : quietScans + 1;
+  }
+  if (quietScans < 3) {
+    throw new Error("worker processes did not reach a quiescent state");
+  }
+} catch (error) {
+  if (processIdentity(watchdog.pid) === watchdogStart) {
+    try { process.kill(watchdog.pid, "SIGTERM"); } catch (killError) { if (!killError || killError.code !== "ESRCH") throw killError; }
+  }
+  resumeProcesses([...frozen].map(([pid, start]) => ({ pid, start })));
+  try { fs.unlinkSync(leasePath); } catch (unlinkError) { if (!unlinkError || unlinkError.code !== "ENOENT") throw unlinkError; }
+  throw error;
+}
+function watchdogMain(watchedLeasePath, watchedNonce) {
+  const check = () => {
+    try {
+      const watchdogFs = require("node:fs");
+      const lease = JSON.parse(watchdogFs.readFileSync(watchedLeasePath, "utf8"));
+      if (
+        !lease ||
+        lease.version !== 1 ||
+        lease.nonce !== watchedNonce ||
+        !Array.isArray(lease.processes) ||
+        !Number.isSafeInteger(lease.expiresAtMs)
+      ) return;
+      const remainingMs = lease.expiresAtMs - Date.now();
+      if (remainingMs > 0) {
+        setTimeout(check, Math.min(remainingMs, 60 * 1000));
+        return;
+      }
+      // Re-read at expiry so a renewal that raced this wake-up wins before SIGCONT.
+      const latest = JSON.parse(watchdogFs.readFileSync(watchedLeasePath, "utf8"));
+      if (
+        latest &&
+        latest.version === 1 &&
+        latest.nonce === watchedNonce &&
+        Array.isArray(latest.processes) &&
+        Number.isSafeInteger(latest.expiresAtMs) &&
+        latest.expiresAtMs > Date.now()
+      ) {
+        setTimeout(check, Math.min(latest.expiresAtMs - Date.now(), 60 * 1000));
+        return;
+      }
+      for (const entry of lease.processes) {
+        if (
+          !entry ||
+          !Number.isSafeInteger(entry.pid) ||
+          entry.pid < 1 ||
+          typeof entry.start !== "string" ||
+          processIdentity(entry.pid) !== entry.start
+        ) continue;
+        try { process.kill(entry.pid, "SIGCONT"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
+      }
+      watchdogFs.unlinkSync(watchedLeasePath);
+    } catch (error) {
+      if (!error || error.code !== "ENOENT") process.exitCode = 1;
+    }
+  };
+  check();
+}
+process.stdout.write("quiesced " + nonce + "\n");
+`;
+
+export const REMOTE_WORKSPACE_RENEW_QUIESCENCE_JS = String.raw`const childProcess = require("node:child_process");
+const crypto = require("node:crypto");
+const fs = require("node:fs");
+const os = require("node:os");
+const path = require("node:path");
+const root = fs.realpathSync(process.argv[1]);
+const nonce = process.argv[2];
+const timeoutMs = Number(process.argv[3] || 12 * 60 * 1000);
+const validationMode = process.argv[4] || "final";
+if (typeof process.getuid !== "function") throw new Error("workspace quiescence requires POSIX");
+const uid = process.getuid();
+if (!/^[a-f0-9]{32}$/.test(nonce || "")) throw new Error("invalid workspace quiescence nonce");
+if (!Number.isSafeInteger(timeoutMs) || timeoutMs < 10 * 1000) throw new Error("invalid watchdog timeout");
+if (validationMode !== "heartbeat" && validationMode !== "final") throw new Error("invalid workspace quiescence validation mode");
+const leasePath = path.join(os.homedir(), ".openclaw-worker", "quiescence", crypto.createHash("sha256").update(root).digest("hex") + "." + nonce + ".json");
+${REMOTE_QUIESCENCE_PS_JS}
+${REMOTE_QUIESCENCE_LEASE_JS}
+const input = parseLease(fs.readFileSync(leasePath, "utf8"), nonce, {
+  requireWatchdog: true,
+  minimumRemainingMs: 5000,
+  errorMessage: "workspace quiescence lease is no longer active",
+});
+function writeLease(processes, expiresAtMs) {
+  // renewalQueue is the nonce's only writer; the watchdog only reads this lease.
+  persistLease(leasePath, { ...input, processes, expiresAtMs }, (current) => {
+    if (current.nonce !== nonce || current.watchdog?.pid !== input.watchdog.pid || current.watchdog?.start !== input.watchdog.start) {
+      throw new Error("workspace quiescence lease changed during renewal");
+    }
+  });
+}
+function assertWatchdogActive() {
+  const status = processStatus(input.watchdog.pid);
+  if (!status || status.start !== input.watchdog.start) {
+    throw new Error("workspace quiescence watchdog identity changed unexpectedly");
+  }
+  try { process.kill(input.watchdog.pid, 0); } catch (error) {
+    if (error && error.code === "ESRCH") throw new Error("workspace quiescence watchdog exited unexpectedly");
+    throw error;
+  }
+}
+function refreshLease(processes) {
+  assertWatchdogActive();
+  input.expiresAtMs = Date.now() + timeoutMs;
+  writeLease(processes, input.expiresAtMs);
+}
+for (const entry of input.processes) {
+  const status = processStatus(entry.pid);
+  if (!status || status.start !== entry.start) continue;
+  if (status.state && !status.state.startsWith("T")) throw new Error("workspace quiescence process resumed unexpectedly");
+}
+refreshLease(input.processes);
+if (validationMode === "final") {
+  const frozen = new Map(input.processes.map((entry) => [entry.pid, entry.start]));
+  let quietScans = 0;
+  const sleeper = new Int32Array(new SharedArrayBuffer(4));
+  // A control tunnel can reconnect after the initial freeze; enroll every late process.
+  for (let attempt = 0; attempt < 250 && quietScans < 3; attempt += 1) {
+    const candidates = quiescenceCandidates(
+      processes(),
+      uid,
+      new Set([input.watchdog.pid]),
+    );
+    if (candidates.length + frozen.size > 4096) {
+      throw new Error("too many worker processes to quiesce safely");
+    }
+    for (const [pid, row] of candidates) frozen.set(pid, row.start);
+    let frozenEntries = [...frozen].map(([pid, start]) => ({ pid, start }));
+    refreshLease(frozenEntries);
+    for (const [pid, row] of candidates) {
+      try {
+        if (input.expiresAtMs - Date.now() < 5000) refreshLease(frozenEntries);
+        const current = processStatus(pid);
+        if (!current || current.start !== row.start) {
+          frozen.delete(pid);
+          continue;
+        }
+        if (input.expiresAtMs - Date.now() < 2500) refreshLease(frozenEntries);
+        process.kill(pid, "SIGSTOP");
+      } catch (error) {
+        if (!error || error.code !== "ESRCH") throw error;
+        frozen.delete(pid);
+      }
+    }
+    frozenEntries = [...frozen].map(([pid, start]) => ({ pid, start }));
+    refreshLease(frozenEntries);
+    Atomics.wait(sleeper, 0, 0, 20);
+    const unknownProcess = quiescenceCandidates(
+      processes(),
+      uid,
+      new Set([input.watchdog.pid]),
+    ).length > 0;
+    quietScans = candidates.length > 0 || unknownProcess ? 0 : quietScans + 1;
+  }
+  if (quietScans < 3) {
+    throw new Error("worker processes did not return to a quiescent state");
+  }
+  input.processes = [...frozen].map(([pid, start]) => ({ pid, start }));
+}
+const renewed = { ...input, expiresAtMs: Date.now() + timeoutMs };
+refreshLease(renewed.processes);
+renewed.expiresAtMs = input.expiresAtMs;
+const confirmed = JSON.parse(fs.readFileSync(leasePath, "utf8"));
+if (confirmed.nonce !== nonce || confirmed.expiresAtMs !== renewed.expiresAtMs) {
+  throw new Error("workspace quiescence renewal was not durable");
+}
+process.stdout.write("renewed " + nonce + "\n");
+`;
+
+export const REMOTE_WORKSPACE_RESUME_JS = String.raw`const childProcess = require("node:child_process");
+const crypto = require("node:crypto");
+const fs = require("node:fs");
+const os = require("node:os");
+const path = require("node:path");
+if (typeof process.getuid !== "function") throw new Error("workspace quiescence requires POSIX");
+const root = fs.realpathSync(process.argv[1]);
+const nonce = process.argv[2];
+if (!/^[a-f0-9]{32}$/.test(nonce || "")) throw new Error("invalid workspace quiescence nonce");
+const leasePath = path.join(os.homedir(), ".openclaw-worker", "quiescence", crypto.createHash("sha256").update(root).digest("hex") + "." + nonce + ".json");
+let raw;
+try { raw = fs.readFileSync(leasePath, "utf8"); } catch (error) {
+  if (error && error.code === "ENOENT") process.exit(0);
+  throw error;
+}
+${REMOTE_QUIESCENCE_PS_JS}
+${REMOTE_QUIESCENCE_LEASE_JS}
+const input = parseLease(raw, nonce);
+if (input.watchdog !== null && processIdentity(input.watchdog.pid) === input.watchdog.start) {
+  try { process.kill(input.watchdog.pid, "SIGTERM"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
+}
+for (const entry of input.processes) {
+  if (processIdentity(entry.pid) !== entry.start) continue;
+  try { process.kill(entry.pid, "SIGCONT"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
+}
+fs.unlinkSync(leasePath);
+`;
