+++ src/gateway/worker-environments/workspace-sync-scripts.ts
@@ -11,303 +11,6 @@ import {
 } from "./workspace-path-exclusions.js";
 export { REMOTE_WORKSPACE_SETUP_SCRIPT } from "./workspace-sync-setup-script.js";
 
-export const REMOTE_WORKSPACE_QUIESCE_JS = String.raw`const childProcess = require("node:child_process");
-const crypto = require("node:crypto");
-const fs = require("node:fs");
-const os = require("node:os");
-const path = require("node:path");
-const root = fs.realpathSync(process.argv[1]);
-if (typeof process.getuid !== "function") throw new Error("workspace quiescence requires POSIX");
-const uid = process.getuid();
-if (uid === 0) throw new Error("workspace quiescence refuses root-owned worker sessions");
-const sleeper = new Int32Array(new SharedArrayBuffer(4));
-const leaseDirectory = path.join(os.homedir(), ".openclaw-worker", "quiescence");
-fs.mkdirSync(leaseDirectory, { recursive: true, mode: 0o700 });
-fs.chmodSync(leaseDirectory, 0o700);
-const workspaceKey = crypto.createHash("sha256").update(root).digest("hex");
-const nonce = crypto.randomBytes(16).toString("hex");
-const leasePath = path.join(leaseDirectory, workspaceKey + "." + nonce + ".json");
-const watchdogTimeoutMs = Number(process.argv[2] || 12 * 60 * 1000);
-if (!Number.isSafeInteger(watchdogTimeoutMs) || watchdogTimeoutMs < 1) throw new Error("invalid watchdog timeout");
-function processes() {
-  const output = childProcess.execFileSync("ps", ["-axo", "pid=,ppid=,uid=,stat=,lstart="], {
-    encoding: "utf8",
-    maxBuffer: 4 * 1024 * 1024,
-  });
-  const rows = new Map();
-  for (const line of output.split("\n")) {
-    const match = line.trim().match(/^(\d+)\s+(\d+)\s+(\d+)\s+(\S+)\s+(.+)$/);
-    if (!match) continue;
-    rows.set(Number(match[1]), {
-      ppid: Number(match[2]),
-      uid: Number(match[3]),
-      state: match[4],
-      start: match[5],
-    });
-  }
-  return rows;
-}
-function ancestors(rows) {
-  const result = new Set();
-  let pid = process.pid;
-  while (pid > 0 && !result.has(pid)) {
-    result.add(pid);
-    pid = rows.get(pid)?.ppid || 0;
-  }
-  return result;
-}
-const frozen = new Map();
-let watchdogReference = null;
-function processIdentity(pid) {
-  try {
-    const start = childProcess.execFileSync("ps", ["-o", "lstart=", "-p", String(pid)], {
-      encoding: "utf8",
-      maxBuffer: 4096,
-    }).trim();
-    return start || null;
-  } catch (error) {
-    if (error && error.status === 1) return null;
-    throw error;
-  }
-}
-function validProcessReference(value) {
-  return value && Number.isSafeInteger(value.pid) && value.pid > 0 && typeof value.start === "string" && value.start.length > 0 && value.start.length <= 128;
-}
-function parseLease(raw, expectedNonce) {
-  const lease = JSON.parse(raw);
-  if (
-    !lease ||
-    lease.version !== 1 ||
-    lease.nonce !== expectedNonce ||
-    !Array.isArray(lease.processes) ||
-    lease.processes.length > 4096 ||
-    lease.processes.some((entry) => !validProcessReference(entry)) ||
-    (lease.watchdog !== null && !validProcessReference(lease.watchdog)) ||
-    !Number.isSafeInteger(lease.expiresAtMs) ||
-    lease.expiresAtMs < 1
-  ) {
-    throw new Error("invalid workspace quiescence lease");
-  }
-  return lease;
-}
-function persistLease(expiresAtMs = Date.now() + watchdogTimeoutMs) {
-  const temporary = leasePath + "." + process.pid + "." + crypto.randomBytes(8).toString("hex");
-  const processes = [...frozen].map(([pid, start]) => ({ pid, start }));
-  fs.writeFileSync(temporary, JSON.stringify({ version: 1, nonce, processes, watchdog: watchdogReference, expiresAtMs }), { mode: 0o600, flag: "wx" });
-  fs.renameSync(temporary, leasePath);
-}
-function resumeProcesses(entries) {
-  for (const entry of entries) {
-    if (processIdentity(entry.pid) !== entry.start) continue;
-    try {
-      process.kill(entry.pid, "SIGCONT");
-    } catch (error) {
-      if (!error || error.code !== "ESRCH") throw error;
-    }
-  }
-}
-const orphanNames = fs.readdirSync(leaseDirectory).filter((name) =>
-  name.startsWith(workspaceKey + ".") && name.endsWith(".json"),
-);
-if (orphanNames.length > 16) throw new Error("too many workspace quiescence leases");
-for (const name of orphanNames) {
-  const match = name.match(/^[a-f0-9]{64}\.([a-f0-9]{32})\.json$/);
-  if (!match) continue;
-  const orphanPath = path.join(leaseDirectory, name);
-  const lease = parseLease(fs.readFileSync(orphanPath, "utf8"), match[1]);
-  if (lease.watchdog !== null && processIdentity(lease.watchdog.pid) === lease.watchdog.start) {
-    try { process.kill(lease.watchdog.pid, "SIGTERM"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
-  }
-  resumeProcesses(lease.processes);
-  fs.unlinkSync(orphanPath);
-}
-persistLease();
-const watchdog = childProcess.spawn(
-  process.execPath,
-  ["-e", "(" + watchdogMain.toString() + ")(process.argv[1], process.argv[2])", leasePath, nonce],
-  { detached: true, stdio: "ignore" },
-);
-watchdog.unref();
-if (!Number.isSafeInteger(watchdog.pid) || watchdog.pid < 1) {
-  fs.unlinkSync(leasePath);
-  throw new Error("workspace quiescence watchdog did not start");
-}
-let watchdogStart = null;
-for (let attempt = 0; attempt < 100 && !watchdogStart; attempt += 1) {
-  watchdogStart = processIdentity(watchdog.pid);
-  if (!watchdogStart) Atomics.wait(sleeper, 0, 0, 10);
-}
-if (!watchdogStart) {
-  try { process.kill(watchdog.pid, "SIGTERM"); } catch {}
-  fs.unlinkSync(leasePath);
-  throw new Error("workspace quiescence watchdog identity was not observable");
-}
-watchdogReference = { pid: watchdog.pid, start: watchdogStart };
-persistLease();
-let quietScans = 0;
-try {
-  for (let attempt = 0; attempt < 250 && quietScans < 3; attempt += 1) {
-    const before = processes();
-    const preserved = ancestors(before);
-    const candidates = [...before.entries()].filter(
-      ([pid, row]) =>
-        row.uid === uid &&
-        !preserved.has(pid) &&
-        row.ppid !== process.pid &&
-        pid !== watchdog.pid &&
-        !frozen.has(pid) &&
-        !row.state.startsWith("T") &&
-        !row.state.startsWith("Z") &&
-        !row.state.startsWith("X"),
-    );
-    if (candidates.length + frozen.size > 4096) {
-      throw new Error("too many worker processes to quiesce safely");
-    }
-    for (const [pid, row] of candidates) {
-      try {
-        frozen.set(pid, row.start);
-        persistLease();
-        if (processIdentity(pid) !== row.start) {
-          frozen.delete(pid);
-          persistLease();
-          continue;
-        }
-        process.kill(pid, "SIGSTOP");
-      } catch (error) {
-        if (!error || error.code !== "ESRCH") throw error;
-      }
-    }
-    Atomics.wait(sleeper, 0, 0, 20);
-    const after = processes();
-    const afterPreserved = ancestors(after);
-    const writable = [...after.entries()].some(
-      ([pid, row]) =>
-        row.uid === uid &&
-        !afterPreserved.has(pid) &&
-        row.ppid !== process.pid &&
-        pid !== watchdog.pid &&
-        !row.state.startsWith("T") &&
-        !row.state.startsWith("Z") &&
-        !row.state.startsWith("X"),
-    );
-    quietScans = writable ? 0 : quietScans + 1;
-  }
-  if (quietScans < 3) {
-    throw new Error("worker processes did not reach a quiescent state");
-  }
-} catch (error) {
-  if (processIdentity(watchdog.pid) === watchdogStart) {
-    try { process.kill(watchdog.pid, "SIGTERM"); } catch (killError) { if (!killError || killError.code !== "ESRCH") throw killError; }
-  }
-  resumeProcesses([...frozen].map(([pid, start]) => ({ pid, start })));
-  try { fs.unlinkSync(leasePath); } catch (unlinkError) { if (!unlinkError || unlinkError.code !== "ENOENT") throw unlinkError; }
-  throw error;
-}
-function watchdogMain(watchedLeasePath, watchedNonce) {
-  const check = () => {
-    try {
-      const watchdogFs = require("node:fs");
-      const lease = JSON.parse(watchdogFs.readFileSync(watchedLeasePath, "utf8"));
-      if (
-        !lease ||
-        lease.version !== 1 ||
-        lease.nonce !== watchedNonce ||
-        !Array.isArray(lease.processes) ||
-        !Number.isSafeInteger(lease.expiresAtMs)
-      ) return;
-      const remainingMs = lease.expiresAtMs - Date.now();
-      if (remainingMs > 0) {
-        setTimeout(check, Math.min(remainingMs, 60 * 1000));
-        return;
-      }
-      // Re-read at expiry so a renewal that raced this wake-up wins before SIGCONT.
-      const latest = JSON.parse(watchdogFs.readFileSync(watchedLeasePath, "utf8"));
-      if (
-        latest &&
-        latest.version === 1 &&
-        latest.nonce === watchedNonce &&
-        Array.isArray(latest.processes) &&
-        Number.isSafeInteger(latest.expiresAtMs) &&
-        latest.expiresAtMs > Date.now()
-      ) {
-        setTimeout(check, Math.min(latest.expiresAtMs - Date.now(), 60 * 1000));
-        return;
-      }
-      const watchdogChildProcess = require("node:child_process");
-      const identity = (pid) => {
-        try {
-          return watchdogChildProcess.execFileSync("ps", ["-o", "lstart=", "-p", String(pid)], { encoding: "utf8", maxBuffer: 4096 }).trim() || null;
-        } catch (error) {
-          if (error && error.status === 1) return null;
-          throw error;
-        }
-      };
-      for (const entry of lease.processes) {
-        if (
-          !entry ||
-          !Number.isSafeInteger(entry.pid) ||
-          entry.pid < 1 ||
-          typeof entry.start !== "string" ||
-          identity(entry.pid) !== entry.start
-        ) continue;
-        try { process.kill(entry.pid, "SIGCONT"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
-      }
-      watchdogFs.unlinkSync(watchedLeasePath);
-    } catch (error) {
-      if (!error || error.code !== "ENOENT") process.exitCode = 1;
-    }
-  };
-  check();
-}
-process.stdout.write("quiesced " + nonce + "\n");
-`;
-
-export const REMOTE_WORKSPACE_RESUME_JS = String.raw`const crypto = require("node:crypto");
-const fs = require("node:fs");
-const os = require("node:os");
-const path = require("node:path");
-if (typeof process.getuid !== "function") throw new Error("workspace quiescence requires POSIX");
-const root = fs.realpathSync(process.argv[1]);
-const nonce = process.argv[2];
-if (!/^[a-f0-9]{32}$/.test(nonce || "")) throw new Error("invalid workspace quiescence nonce");
-const leasePath = path.join(os.homedir(), ".openclaw-worker", "quiescence", crypto.createHash("sha256").update(root).digest("hex") + "." + nonce + ".json");
-let raw;
-try { raw = fs.readFileSync(leasePath, "utf8"); } catch (error) {
-  if (error && error.code === "ENOENT") process.exit(0);
-  throw error;
-}
-const input = JSON.parse(raw);
-if (
-  !input ||
-  input.version !== 1 ||
-  input.nonce !== nonce ||
-  !Array.isArray(input.processes) ||
-  input.processes.length > 4096 ||
-  input.processes.some((entry) => !entry || !Number.isSafeInteger(entry.pid) || entry.pid < 1 || typeof entry.start !== "string" || !entry.start || entry.start.length > 128) ||
-  (input.watchdog !== null && (!input.watchdog || !Number.isSafeInteger(input.watchdog.pid) || input.watchdog.pid < 1 || typeof input.watchdog.start !== "string" || !input.watchdog.start || input.watchdog.start.length > 128)) ||
-  !Number.isSafeInteger(input.expiresAtMs) ||
-  input.expiresAtMs < 1
-) {
-  throw new Error("invalid workspace quiescence lease");
-}
-function identity(pid) {
-  try {
-    return require("node:child_process").execFileSync("ps", ["-o", "lstart=", "-p", String(pid)], { encoding: "utf8", maxBuffer: 4096 }).trim() || null;
-  } catch (error) {
-    if (error && error.status === 1) return null;
-    throw error;
-  }
-}
-if (input.watchdog !== null && identity(input.watchdog.pid) === input.watchdog.start) {
-  try { process.kill(input.watchdog.pid, "SIGTERM"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
-}
-for (const entry of input.processes) {
-  if (identity(entry.pid) !== entry.start) continue;
-  try { process.kill(entry.pid, "SIGCONT"); } catch (error) { if (!error || error.code !== "ESRCH") throw error; }
-}
-fs.unlinkSync(leasePath);
-`;
-
 export const REMOTE_GIT_WORKSPACE_SETUP_SCRIPT = String.raw`set -eu
 workspace=$1
 pack=$2
@@ -369,7 +72,6 @@ const requestedBaseCommit = process.argv[2] || null;
 const eligibleOnly = process.argv[3] === "eligible";
 const requestedManifestDigest = process.argv[3] === "resolve" ? process.argv[4] : null;
 const publishedManifestDigest = process.argv[3] === "publish" ? process.argv[4] : null;
-const legacyGatewayLocale = requestedManifestDigest ? process.argv[5] : null;
 const priorManifestDigests = [...new Set(process.argv.slice(4).filter(Boolean))];
 const entriesByPath = new Map();
 function fail(message) {
